requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
Markdown==2.6.8
markdown2==2.3.3
Pillow==4.0.0
isort==4.2.5
flake8==3.3.0